backend/requirements.txt
fastapi==0.128.0
uvicorn==0.37.0
sqlalchemy==2.0.46
pydantic==2.12.2
python-dotenv==1.1.0
email-validator==2.3.0
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
